Fix *_workers config settings.
Admin_ and public_workers were in the wrong section. They belong to the
evenlet_server section and only impact the eventlet configuration. In
addition, they are deprecated and anyone setting them will get warnings
in their logs. For now we should mark them deprecated and get them in
the right section, but once eventlet is removed, they will also be
removed.

Change-Id: Ide4b80f0b7794ecd81599c28f65326488c8d6ed6