Deletes tenant parameter from keystone_user
The tenant parameter was made deprecated in Kilo and to be removed in Liberty:
see bug/1472437 and corresponding patch.

This patch removes the tenant parameter from the keystone_user provider and the
manifests using it.

From now on, a user is assigned to a project (tenant) through the
keystone_user_role provider.

For instance,

When using the default domain:
  keystone_user_role {'user1@project':
    ensure => present,
    roles  => ['admin'],
  }

When using non default domains:
  keystone_user_role {'user1::domain1@project1::domain2':
    ensure => present,
    roles  => ['admin'],
  }

Change-Id: I79f1bca4c67a4c92234431b2ae867933172a1ebe