Support for services different only by type.
First, Keystone_service provider was not able to create service
different by type.  So service type computev3 and compute with name nova
could not be created.

Second, on a system with such resource it would not be able to
differentiate between the two.  Then if

   keystone_service {'nova': type => 'compute'}

was in the catalog, then the service nova/computev3 would be changed to
nova/compute.

Now you can specify:

   keystone_service {'service::type'}

or

   keystone_service {'service': type => 'type'}

It keeps a full backward compatibility.

Change-Id: I24bbcf8cbf3760fd061e2439f37864ae97ce86f6
Closes-Bug: #1508886