Keystone_endpoint match service by name/type.
This enable keystone_endpoint to specify the type of the service
matched.  This way one can match services which are different only by
type and not only by name, like services nova/compute and nova/computev3
for instance.  It does so by fetching the _id_ of the service when it
has the type information instead of just using the name.

This should be required, and deprecation has been added, as the current
code work only because of a convention.

Change-Id: I9ea20fbad274d583485bc09a52b9df8000eb1af5
Closes-Bug: #1506996