Remove references to puppet-openstack from README
The README references to the retired puppet-openstack module
should be removed, as they link to the empty github page.

This patch also cleans up a few other parts of the README,
such as adding a link to the keystone wiki and removing
'%example' from the installation command to match the
other projects.

Change-Id: Ie74ca8a1914cbb2d3de199001b33ced181501a45
Closes-bug: #1518589