Adjust rspec tests for retries in openstacklib.
A patch in openstacklib[1] broke the some rspec test in keystone.  This
adjust the test to match the new behavior of the lib.

The request_timeout is fixed to 0 so that only one retry is done.
Failure to do that would make the number of retries unknown and the time
spend on the three failing tests around 60 seconds each.

[1] https://github.com/openstack/puppet-openstacklib/commit/701b6fe0acf56e9132e9b0a448f28670c714cedb

Change-Id: I6adc439a806cf209e2fae35446a15b07353a18c2
Closes-Bug: #1529371