Fix cache and cache memcache configurations
This change adds the ability to manage the cache memcache servers.
Previously we were enabling the cache if memcache servers were set but
improperly configuring the cache configuration options for keystone.
The memcache servers were being set for memcache/servers but
the cache configuration looks at cache/memcache_servers for its server
list. With this change we are adding a cache_memcache_servers parameter
to allow the configuring the cache/memcache_servers. We are continuing to
use the memcache_servers as the default if cache_memcache_servers is not
provided.

Additionally we are allowing the manual enabling and disabling of the
cache configuration but will fall back to the previous behaviour
of enabling the cache if memcache servers are specified.

New parameters are:
  - cache_memcache_servers
  - cache_enabled

Change-Id: I6f86c7f8f55a6f7a7e8caa922c55e618fec8e392
Closes-Bug: #1523393