endpoint: stop sending warning if $version is not empty
No all OpenStack clients & Tempest[1] actually support versionless
endpoints, that's why we need to wait more before changing the default
value.

This patch aims to:
* document why we keep to set a version by default
* stop to show a warning when not setting empty.

Setting the $version to empty is at the risk of the deployer, and note
that it could lead to issues with some OpenStack clients & Tempest.

[1] https://review.openstack.org/#/q/status:open+project:openstack/tempest-lib+branch:master+topic:bug/1530181

Change-Id: Ie2c8070411d403bfcd372665462922db077ee961