Specify user and project domains for admin role
In keystone::roles::admin, admin_project_domain and admin_user_domain
are not applied to the admin role. This results in errors when
applying the role as it uses the "Default" domain in
keystone_user_role:

  class { '::keystone::roles::admin':
    email => 'marcus@aptira.com',
    password => $admin_password,
    admin => 'admin', # username
    admin_tenant => 'admin', # project name
    admin_user_domain => 'admin', # domain for user
    admin_project_domain => 'admin', # domain for project
  }

Error: /Stage[main]/Keystone::Roles::Admin/Keystone_user_role[admin@admin]: Could not evaluate: No project admin with domain Default found

This patch adds the admin_project_domain and the admin_user_domain to the
role.

Change-Id: Ia3f899dfb78b0887f31ee82d6b21d2fb2536ad84
Closes-Bug: #1533913