Fix rabbitmq ssl logic
Saying ! is_service_default($rabbit_use_ssl) and $rabbit_use_ssl is
invalid logic: it checks that the value is not the ensure=>absent
default and is also boolean true, when it should be checking whether it
is boolean false. This was leading to an error where if the kombu_*
parameters were not set but $rabbitmq_use_ssl was set to true, puppet
would fail. It is valid to set rabbitmq_use_false to true while not
setting the kombu_* parameters, it is just not valid the other way
around. This change fixes the logic, adds checks for definition and
truthiness of the kombu_* params, and fixes the error message. This
makes the logic identical to the rabbitmq SSL handling in the neutron
module[1].

[1] http://git.openstack.org/cgit/openstack/puppet-neutron/tree/manifests/init.pp#n356

Closes-Bug: #1356083
Change-Id: I1f4a67c4b8d43670793331254818b6991b8de640