Do not manage keystone provider resources if prefetch fails
If prefetch of Keystone provider resources fails (because of
Keystone API timeouts or failures, for example) then the provider
will believe the resources don't exist and will attempt to create
them.  This can cause duplicate resources in Keystone.

This patch records when the prefetch fails, and if so, shunts any
of the resource management methods to fail().

Based on the similar patch to puppet-neutron:
https://review.openstack.org/#/c/265386/2

Change-Id: I8f41c8a8594c55b9bc0931f6d94461190745ad1b
Closes-bug: 1546357