Fixed require statements for puppet 4
This change updates the require statements in the providers
to resolve the issues with the autoloader in puppet. This
issue technically always existed but was masked prior to the
split with puppet server and puppet agent because both used to
share the same loadpath. Documentation on this issue can be
read at https://tickets.puppetlabs.com/browse/PUP-4450.

Change-Id: I5eae546d65832a39dd75ab9c0f97868c7edc9e91