Fix ldap_backend and puppet 4.6
On xenial after update from puppet-agent 1.5 to 1.6, which include an
upgrade from 4.5 to 4.6 of puppet this code stopped working.

The ensure_resource in the keystone/init.pp manifest was not evaluated.
Adding this include seems to make it work again and we see that is the
log:

```
     Debug: Resource keystone_config[identity/domain_specific_drivers_enabled] was not determined to be defined
     Debug: Create new resource keystone_config[identity/domain_specific_drivers_enabled] with params {"value"=>true}
```

while without the include ::keystone, it's not there.

Puppet guru needed as to the why.

Change-Id: Ief78d70b8fe114ddf40d98fab93374862d3c23cb