Add support for credential_setup
Allow and enable by default the setup of Keystone credentials with keystone-manage.
It has been a requirement in Keystone upstream so puppet-keystone will support the
management of credential directory, keystone-manage credential_setup execution (can
be disabled with enable_credential_setup boolean) and the configuration of
credential/key_repository in keystone.conf.

Note: the feature is now disabled by default because UCA doesn't have
latest keystone. Once it has it, the value will be switched to True.

Change-Id: I364f65c7009e400f2851e37999e1d8c0b6f5fd5a