Add admin_password parameter for use in bootstrap
keystone-manage bootstrap should be taking the admin_password, not the
admin token. The admin_token is deprecated in Keystone and will be
deprecated in this module in a later release. For now we add an
admin_password parameter which defaults to admin_token, but do not rely
on the default value since admin_token will go away. Beyond the
incorrect usage of admin_token that this fixes, admin_token is still
used in other places and that is why it will be removed in a separate
change.

Closes-Bug: #1621959

Change-Id: I7a706d93b43ec025bdb4b29667f64ff2f7dd52a0