Allow the management of the Fernet Keys
When keystone uses Fernet tokens in a multi node
environment, each of the nodes needs to have a
copy of the keys used by other nodes.  While the
upstream suggestion is to run the command on one
node and then copy the keys to the others,
deployments using a central configuration server
do not usually follow workflows like this. Instead
the keys are generated off server, and managed in a
central data store.

This change follows the pattern set for credentials
key management.

Change-Id: Ibd2a7692c247d0367c5fd331bb88790f882c2c91