Accept null value for servername_admin in apache resource
The servername_admin parameter uses the servername itself to get a
default value. This parameter, is based on the fqdn fact, which can
be null. However, servername_admin is set with the pick function
which doesn't take null values, so it crashes when that's the case.
This takes pick_default into use which actually takes null values and
fixes the issue.

Change-Id: Id8f3e4467b7081802cd2e92414679c49953fcd02
Closes-Bug: #1625010