Implement crontab to perform Fernet keys rotations
This crontab is useful to run keystone-manage fernet_rotate command in a
scheduled way.
It doesn't take care of the distribution of keys when deploying Keystone
on multinode environment but it's still useful to use it on single-node
keystone deployments or when we have an external system to distribute
the keys after running this command.

Change-Id: I125e81d8cd130fadb8271f1b7bcdcf9794c79f47