Change keystone token flush to run hourly
In a recent commit [1] the keystone token flush cron job was changed to
run twice a day. However, this change was not enough for big
deployments.

After getting some customer feedback and looking at what other projects
are doing [2] [3] [4]. It seems that running this job hourly is the way
to go.

[1] Ia0b0fb422318712f4b0f4d023cbb3a61d40bb85d
[2] https://www.ibm.com/support/knowledgecenter/en/SSB27U_6.4.0/com.ibm.zvm.v640.hcpo4/exptoken.htm
[3] https://review.openstack.org/#/c/88670/8
[4] https://github.com/openstack/charm-keystone/blob/master/templates/keystone-token-flush

Change-Id: I6ec7ec8111bd93e5638cfe96189e36f0e0691d65
Related-Bug: #1649616