Make sure that token flush's  maxdelay is a numeric value
If this is given something else than a number, it will add an erroneous
value to the cron job and it'll fail.

The original commit added the Integer type to the maxdelay, however
due to feedback and the fact that we can't asure that folks will be
using puppet 4 with stable/ocata. I added a legacy validation function.

Depends-On: I00823e23358df91ce54f421c12636f05d4196e15
Change-Id: Iccb1958d0237ec4105c30e08db86e28cc630d08b
Closes-Bug: #1708584