Puppet class to disable the v2.0 API for Keystone through keystone-paste.ini
This class removes the /v2.0 API entries for both the public and admin API
endpoints, ensuring that the only surviving API is v3. Currently we have both
enabled, as there may be some legacy reasons to keep the v2.0 API enabled,
but some may want an explicit way to disable the older v2.0 API. It's intended
that this class would be optionally called when required, e.g. via TripleO.

Change-Id: I635cbda7adf33eb47dca81b4bab5c8644bf5b3d8