Make it so that service_name can handle Debian
If the value is 'keystone-public-keystone-admin', then the
module will use 2 services, one called keystone-public, and
one called keystone-admin (as per the new Debian package
which uses UWSGI instead of Apache).

Note: the Debian package doesn't do like this *yet*, but
that is planned for when this puppet module will support it.

Change-Id: I0e6772dc22e8c27a2ffc767844c7eae12e521235