authtoken cannot manage python-memcache on RedHat
The keystone::resource::authtoken define does not
have access to the keystone::params class so it
used the package resource title as package name
when trying to install.

This resulted in the non-existing python-memcache
package failed to install on RedHat based systems.

This includes the params class inside the define
so it gets access to the python_memcache_package_name
variable.

Change-Id: I3470292d87620db717251092fbacf16b2cace571
Closes-Bug: 1711437