Add manage_backend_package parameter
Adds the manage_backend_package parameter that is sent
to the oslo::cache class that determines if we should
install the backend python library for caching.

The default value is not changed and defaults to true
same as the oslo::cache class does.

Change-Id: I8a99eee7622e25847c2d729fa97d0500fe27d8f4
(cherry picked from commit f882e39447506eb500e3211df786c7d21b84d792)