Remove deprecated keystone::memcache_servers
The memcache_servers has been deprecated a long time and should have
been removed in Newton and the memcache/servers option in keystone.conf
does not exist anymore.

This removes the memcache_servers parameter and removed the behavior
that automatically changed cache_backend and cache_enabled when that
option was specified.

Users must now explicitly set cache_backend, cache_enabled and
cache_memcache_servers to enable the presence of caching.

Change-Id: I3f5a4acd76b312ada2ae65dd10e4a426cb22c131