Correct openid federation protocol name in httpd template
Keystone requires the auth method for federation using OpenID
Connect to be 'openid', which we already enforce in puppet-keystone.
The protocol name that shows up in the federation URLs in the httpd
config must match the auth method.  We have a few locations that
don't match, which results in an invalid configuration.

This patch corrects the protocol in the locations protected by
mod_auth_openidc and the OIDCRedirectURI directives.

Change-Id: I02027676c532b447e56c8774c25015ec9a40b414