Fix keystone::client testing
Fixes the failing testing since the package resource in the
openstacklib::openstackclient class was changed [1].

Modules testing should *never* check names of other resources
but only the actual resources it's using, which in this case is
the openstacklib::openstackclient class and not the package.

Also takes the oppertunity to convert the testing to using
rspec-puppet-facts, the keystone modules is the only module
that is left until all modules are using rspec-puppet-facts.

[1] http://logs.openstack.org/59/632459/9/check/puppet-openstack-unit-4.8-centos-7/0886361/job-output.txt.gz#_2019-01-28_13_16_02_682694

Change-Id: I4c86797fba211387a53036ce933e430daefb8f0e