Fix database test in keystone class test
We were checking for the value of a database setting which is now
managed via puppet-olso in the keystone init class.  We should not be
checking the values set by another class in the unit tests. This change
updates the unit tests to just make sure we're including the
keystone::db class rather than the values configured by it.

Change-Id: I0a1b1adec24ba528b623ad0b17e5bba16e0b279a