Fix performance regression due to reduced number of keystone workers
The puppet-keystone project consolidated keystone applications:

  https://github.com/openstack/puppet-keystone/commit/ace7aeb3b71b39a59f92fbc9e7f676a70c9a797a#diff-e5968f40345cf5d3be0539fbba87f787

This effectively reduces the number of available workers to process
keystone requests since only half the processes are deployed now that
the applications are consolidated [0]. As a result, we see a performance
regression [1] when interacting with keystone.

For example, in earlier versions, an 8 core machine would spawn four
processes to serve the keystone-main application. Another four processes
would be spawned to serve the keystone-admin application. Now, we only
have 4 processes total to serve the entire keystone application. Due to
the difference in the number of workers deployed by default, users will
experience reduced performance out-of-the-box.

This patch not only brings performance back to the previous levels, it
is also safe to implement as we are not changing the number of workers
when compared to previous releases and only bringing parity.

[0] https://github.com/openstack/puppet-openstacklib/blob/master/lib/facter/os_workers.rb#L33
[1] https://bugzilla.redhat.com/show_bug.cgi?id=1789495

Co-Authored-By: Takashi Kajinami <tkajinam@redhat.com>
Depends-on: https://review.opendev.org/#/c/705041/
Change-Id: Icf9acaa106af705fa249a2ef2abca9f9a91fba59