Add support for JWKS based OAuth Token validation.
Keycloak only enables the introspection API when using a 'confidential'
client.  This means that for user initiated OAuth2 tokens the user would
need to know the secrets for a client (ie the confidential piece).

With Keycloak to use a 'public' client, where the user doesn't need to
know a client secret, you need to use a JWKS to validate the JWT tokens
instead of the introspection endpoint.  OIDCOAuthVerifyJwksUri tells
mod_auth_openidc where to find that certificate.  You also need to
*remove* OIDCOAuthIntrospectionEndpoint as this conflicts with
OIDCOAuthVerifyJwksUri

Change-Id: I95220db87f61ca19e50c64b4f799d3973b256858