Always pass --name when flushing keystone_service
Running "openstack service set <id> --description="
will throw an error [1].

The error occurs because the request to Keystone must include any
of enabled, name or type to satisfy the minProperties of the patch
request.

This patch always sets the --name parameter when calling the openstack
command so that we satisfy this requirement.

[1] http://paste.openstack.org/show/795869/

Change-Id: Ie4e2fe62ee7734f657f765c352ba542360c0e1d7