Revert "Do not set public_bind_host and public_port in eventlet section"
When public_endpoint is set but different urls are used for endpoints
(especially for admin endpoint and public endpoint), it can cause
problem with self-url detection in keystone because it always assumes
that the url should be directed to that public_endpoint even when
a request comes from admin endpoint.

This patch reverts previous change which made public_endpoint parameter
generated automatically if unset, so that public_endpoint is not set
by default.

Note 1: We can't directly remove public_endpoint by default because
this parameter is mandatory for providers in stable/train.
Thus we need to revert the previous change to unset public_bind_host
and public_port so that these 2 parameters can be refered from proviers
instead of public_endpoint, in the deployments which can't depend on
public_endpoint.

This reverts commit d58fcfe75e887af3a1e6f6e9aa9e69752fec5a18.

Related-bug: #1889017
Change-Id: I4feb0dfb858bcf2ff5d97341908abd11f6939b32