Make service desc in keystone::endpoint configurable
I have a issue when upgrading a Train deployment to
Ussuri because the migration to keystone::bootstrap
did not cover the fact that the service description
was set on the keystone_service.

This means that if you migrate to using
keystone::bootstrap it will ensure_resource() on
the keystone_service which defaults the desc to an
empty string causing it to not be idempotent.

This can be avoided by changing the service desc
to match the keystone::bootstrap behavior to a empty
string but this cannot be done since this is hardcoded.

The alternative to this is to change the behavior in
ussuri and master which would instead break deployments
that has worked around that issue already.

Change-Id: I307583b0f88dbedf9aa4fa02eda61533fd863c4b